add upgrade path for a new 4.11.2 systemvmtemplate and support for XS 7.5
